feat: blog  functionality improvments
- fixed  the main /blog page to list all articles.
